24th July
I have added various features for the user experience of my calendar. User has the ability to choose which department of tasks they wish to view.
I added another feature where user can drag and drop their events to a different place on the calendar and this automatically updates the database and saves their task to that place.
Finally, I added different colours for each department for the tasks. This means it is more recognisable for user to tell which task is assigned to each department.